Corrected build errors
